% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{deleteEmptyRows}
\alias{deleteEmptyRows}
\title{deleteEmptyRow}
\usage{
deleteEmptyRows(df)
}
\arguments{
\item{df}{- dataframe with data}
}
\value{
compacted dataframe
}
\description{
Compacts the dataframe and returns only complete rows
}
\examples{
df <- deleteEmptyRows(df)
}

