% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{initialize_dataframe}
\alias{initialize_dataframe}
\title{initialize_dataframe}
\usage{
initialize_dataframe(...)
}
\arguments{
\item{file1, }{file2, .... Objects with xml of the files to process}
}
\value{
empty dataframe with the maximum number of CPEs
}
\description{
Generates a empty dataframe with the size
 of all the CPEs of the file
}
\examples{
df <- initialize_dataframe(xmlFile, xmlFile2)
}

