% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseCVEFile.R
\name{obtainAllCPEs}
\alias{obtainAllCPEs}
\title{obtainAllCPEs}
\usage{
obtainAllCPEs(doc)
}
\arguments{
\item{doc}{- object with loaded file}
}
\value{
CPEs list's file
}
\description{
From the object that contains the CVEs xml file
you get a list with all the CPEs in the file
}
\examples{
list <- obtainAllCPEs(doc)
}

