% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseCVEFile.R
\name{obtainCPEbyCVE}
\alias{obtainCPEbyCVE}
\title{obtainCPEbyCVE}
\usage{
obtainCPEbyCVE(doc, cve)
}
\arguments{
\item{doc}{- object with loaded file}

\item{cve}{- CVE to obtain CPEs}
}
\value{
CPEs list
}
\description{
From the object that contains the CVEs xml file and a CVE code
you get a list with all the CPEs associated
}
\examples{
list <- obtainCPEbyCVE(doc,"CVE-2016-002")
}

