% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseCVEFile.R
\name{obtainCVEs}
\alias{obtainCVEs}
\title{obtainCVEs}
\usage{
obtainCVEs(doc)
}
\arguments{
\item{doc}{- object with loaded file}
}
\value{
CVEs list file
}
\description{
From the object that contains the CVEs xml file
you get a list with all the CVEs in the file
}
\examples{
list <- obtainCVEs(doc)
}

