% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseCVEFile.R
\name{obtainCVSSbyCVE}
\alias{obtainCVSSbyCVE}
\title{obtainCVSSbyCVE}
\usage{
obtainCVSSbyCVE(doc, cve)
}
\arguments{
\item{doc}{- object with loaded file}

\item{cve}{- CVE to obtain CVSSs}
}
\value{
CVSSs list
}
\description{
From the object that contains the CVEs xml file and a CVE code
you get a list with all the CVSSs associated
}
\examples{
list <- obtainCVSSbyCVE(doc,"CVE-2016-002")
}

