% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{obtainLastElement}
\alias{obtainLastElement}
\title{obtainLastElement}
\usage{
obtainLastElement(df)
}
\arguments{
\item{df}{- dataframe}
}
\value{
first empty position
}
\description{
Obtains the first empty position in the dataframe
}
\examples{
pos <- obtainLastElement(df)
}

