% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{save_content}
\alias{save_content}
\title{save_content}
\usage{
save_content(xmlFile, nodesList, df, cpe)
}
\arguments{
\item{xmlFile}{- xml de los CVE's}

\item{df}{- dataframe}

\item{cpe}{- OS to search}

\item{listNodes}{- xml separated like a list}
}
\value{
Dataframe with OS information
}
\description{
Insert the CVEs, CPEs and CVSSs for
the inserted OS
}
\examples{
df <- save_content(xmlFile, listaNodo,df,"android")
}

